fix: create-vesting deploy issues
