+++ packages/cow-hooks-ui/package.json
@@ -21,6 +21,7 @@
     "@types/react": "^18.3.8",
     "@types/react-dom": "^18.3.0",
     "autoprefixer": "^10.4.20",
+    "next": "15.0.0-rc.0",
     "postcss": "^8.4.47",
     "tailwindcss": "^3.4.13",
     "typescript": "5.6.2"
